database/practice/데일리 실습_db_08_3_P/day8_problem3/requirements.txt
asgiref==3.3.4
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
